timeStep passed up to java. screen dimensions passed down to c++.

+++ src/se/abjorklund/math/Vector2.java
@@ -1,27 +1,27 @@
 package se.abjorklund.math;
 
 public class Vector2 {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Vector2(int x, int y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
